requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
asgiref==3.11.0
    # via django
colorama==0.4.6
    # via pytest
django==6.0
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.16.1
    # via -r requirements.in
flake8==7.3.0
    # via -r requirements.in
iniconfig==2.3.0
    # via pytest
mccabe==0.7.0
    # via flake8
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via pytest
pycodestyle==2.14.0
    # via flake8
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via pytest
pytest==9.0.1
    # via -r requirements.in
sqlparse==0.5.4
    # via django
tzdata==2025.2
    # via django
